requirements.txt
cohere
python-dotenv
langchain
annoy
numpy
pandas
gradio
assemblyai
unstructured